요구 사항에 맞춰 변경

+++ src/main/java/com/example/hanghae3rd/service/posts/PostsService.java
@@ -5,9 +5,14 @@ import com.example.hanghae3rd.domain.posts.PostsRepository;
 import com.example.hanghae3rd.web.dto.PostsResponseDto;
 import com.example.hanghae3rd.web.dto.PostsSaveRequestDto;
 import com.example.hanghae3rd.web.dto.PostsUpdateRequestDto;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityTransaction;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -18,6 +23,7 @@ import java.util.List;
 @Service
 public class PostsService {
     private final PostsRepository postsRepository;
+    private final PlatformTransactionManager transactionManager;
 
     //게시글 저장
     @Transactional
@@ -26,23 +32,24 @@ public class PostsService {
     }
 
     //게시글 수정
-    @Transactional
-    public Long update(Long id, PostsUpdateRequestDto requestDto){
+    public void update(Long id, PostsUpdateRequestDto requestDto){
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
 
-        posts.update(requestDto.getTitle(), requestDto.getContent());
-
-        return id;
+        if(posts.getPassword().equals(requestDto.getPassword())){
+            posts.update(requestDto.getTitle(), requestDto.getContent(), requestDto.getAuthor());
+        }
     }
 
     //게시글 삭제
     @Transactional
-    public void delete(Long id){
+    public void delete(Long id, PostsUpdateRequestDto requestDto){
         Posts posts = postsRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
 
-        postsRepository.delete(posts);
+        if(posts.getPassword().equals(requestDto.getPassword())){
+            postsRepository.delete(posts);
+        }
     }
 
     //id에 해당하는 게시글 반환
